@@ -12,8 +12,8 @@
 				<?php
 				printf(
 					esc_html__( 'Copyright &#169; %1$s %2$s', 'pulsar' ),
-					date_i18n( 'Y' ),
-					get_bloginfo( 'name' )
+					esc_html( date_i18n( 'Y' ) ),
+					esc_html( get_bloginfo( 'name' ) ),
 				);
 				?>
 			</div>
