@@ -9,6 +9,7 @@ namespace Pulsar\Editor;
 
 use Pulsar\Contracts\Bootable;
 use Pulsar\Tools\Config;
+use \WP_Block_Editor_Context;
 
 /**
  * Block handling.
@@ -25,7 +26,7 @@ class Blocks implements Bootable {
 	 * @access public
 	 * @return void
 	 */
-	public function boot() {
+	public function boot() : void {
 
 		// Load individual block CSS files on demand.
 		add_filter( 'should_load_separate_core_block_assets', '__return_true' );
@@ -46,7 +47,7 @@ class Blocks implements Bootable {
 	 * @access public
 	 * @return void
 	 */
-	public function register_custom_blocks() {
+	public function register_custom_blocks() : void {
 
 		$blocks_directory = get_theme_file_path( '/dist/blocks/' );
 
@@ -65,20 +66,15 @@ class Blocks implements Bootable {
 	/**
 	 * Limit the blocks that are allowed to content editors.
 	 *
-	 * @param bool|string[]           $block_editor_context Array of block type slugs, or boolean to enable/disable all
-	 * @param WP_Block_Editor_Context $editor_context The current block editor context
-	 * @return array
+	 * @param array|bool               $block_editor_context Array of block type slugs, or boolean to enable/disable all
+	 * @param \WP_Block_Editor_Context $editor_context       The current block editor context
+	 * @return array|bool
 	 */
-	public function allowed_blocks( $block_editor_context, $editor_context ) {
+	public function allowed_blocks( array|bool $block_editor_context, \WP_Block_Editor_Context $editor_context ) : array|bool {
 		if ( self::ENABLE_ALL_BLOCKS || empty( $editor_context->post ) ) {
 			return $block_editor_context;
 		}
 
-		$blocks = Config::get( 'allowed-blocks' );
-		if ( ! $blocks ) {
-			return $block_editor_context;
-		}
-
-		return $blocks;
+		return Config::get( 'allowed-blocks' ) ?? $block_editor_context;
 	}
 }
