@@ -14,8 +14,6 @@ class PrimaryWalker extends \Walker_Nav_Menu {
 	/**
 	 * Starts the list before the elements are added.
 	 *
-	 * @since 3.0.0
-	 *
 	 * @see Walker::start_lvl()
 	 *
 	 * @param string   $output Used to append additional content (passed by reference).
@@ -50,7 +48,19 @@ class PrimaryWalker extends \Walker_Nav_Menu {
 		$output .= "{$n}{$indent}<ul$class_names x-cloak>{$n}";
 	}
 
-	public function start_el( &$output, $item, $depth = 0, $args = null, $id = 0 ) {
+	/**
+	 * Starts the element output.
+	 *
+	 * @param string    $output            Used to append additional content (passed by reference).
+	 * @param \WP_Post  $data_object       Menu item data object.
+	 * @param int       $depth             Depth of menu item. Used for padding.
+	 * @param \stdClass $args              An object of wp_nav_menu() arguments.
+	 * @param int       $current_object_id Optional. ID of the current menu item. Default 0.
+	 */
+	public function start_el( &$output, $data_object, $depth = 0, $args = null, $current_object_id = 0 ) {
+		// Restores the more descriptive, specific name for use within this method.
+		$menu_item = $data_object;
+
 		if ( isset( $args->item_spacing ) && 'discard' === $args->item_spacing ) {
 			$t = '';
 			$n = '';
@@ -60,72 +70,61 @@ class PrimaryWalker extends \Walker_Nav_Menu {
 		}
 		$indent = ( $depth ) ? str_repeat( $t, $depth ) : '';
 
-		$classes   = empty( $item->classes ) ? [] : (array) $item->classes;
-		$classes[] = 'menu-item-' . $item->ID;
+		$classes   = empty( $menu_item->classes ) ? [] : (array) $menu_item->classes;
+		$classes[] = 'menu-item-' . $menu_item->ID;
 
 		/**
 		 * Filters the arguments for a single nav menu item.
 		 *
-		 * @since 4.4.0
-		 *
-		 * @param stdClass $args  An object of wp_nav_menu() arguments.
-		 * @param WP_Post  $item  Menu item data object.
-		 * @param int      $depth Depth of menu item. Used for padding.
+		 * @param stdClass  $args      An object of wp_nav_menu() arguments.
+		 * @param \WP_Post  $menu_item Menu item data object.
+		 * @param int       $depth     Depth of menu item. Used for padding.
 		 */
-		$args = apply_filters( 'nav_menu_item_args', $args, $item, $depth );
+		$args = apply_filters( 'nav_menu_item_args', $args, $menu_item, $depth );
 
 		/**
 		 * Filters the CSS classes applied to a menu item's list item element.
 		 *
-		 * @since 3.0.0
-		 * @since 4.1.0 The `$depth` parameter was added.
-		 *
-		 * @param string[] $classes Array of the CSS classes that are applied to the menu item's `<li>` element.
-		 * @param WP_Post  $item    The current menu item.
-		 * @param stdClass $args    An object of wp_nav_menu() arguments.
-		 * @param int      $depth   Depth of menu item. Used for padding.
+		 * @param string[] $classes   Array of the CSS classes that are applied to the menu item's `<li>` element.
+		 * @param WP_Post  $menu_item The current menu item.
+		 * @param stdClass $args      An object of wp_nav_menu() arguments.
+		 * @param int      $depth     Depth of menu item. Used for padding.
 		 */
-		$class_names = implode( ' ', apply_filters( 'nav_menu_css_class', array_filter( $classes ), $item, $args, $depth ) );
+		$class_names = implode( ' ', apply_filters( 'nav_menu_css_class', array_filter( $classes ), $menu_item, $args, $depth ) );
 		$class_names = $class_names ? ' class="' . esc_attr( $class_names ) . '"' : '';
 
 		/**
 		 * Filters the ID applied to a menu item's list item element.
 		 *
-		 * @since 3.0.1
-		 * @since 4.1.0 The `$depth` parameter was added.
-		 *
-		 * @param string   $menu_id The ID that is applied to the menu item's `<li>` element.
-		 * @param WP_Post  $item    The current menu item.
-		 * @param stdClass $args    An object of wp_nav_menu() arguments.
-		 * @param int      $depth   Depth of menu item. Used for padding.
+		 * @param string   $menu_id   The ID that is applied to the menu item's `<li>` element.
+		 * @param WP_Post  $menu_item The current menu item.
+		 * @param stdClass $args      An object of wp_nav_menu() arguments.
+		 * @param int      $depth     Depth of menu item. Used for padding.
 		 */
-		$id = apply_filters( 'nav_menu_item_id', 'menu-item-' . $item->ID, $item, $args, $depth );
+		$id = apply_filters( 'nav_menu_item_id', 'menu-item-' . $menu_item->ID, $menu_item, $args, $depth );
 		$id = $id ? ' id="' . esc_attr( $id ) . '"' : '';
 
 		$output .= $indent . '<li' . $id . $class_names;
 		if ( $depth === 0 ) {
 			$output .= ' @keydown.escape="onEscape($event)"';
-			$output .= " @pointerleave.debounce.50ms='onPointerLeave(" . $item->ID . ', ' . $item->menu_item_parent . ")'";
+			$output .= " @pointerleave.debounce.50ms='onPointerLeave(" . $menu_item->ID . ', ' . $menu_item->menu_item_parent . ")'";
 
 		}
 		$output .= '>';
 
 		$atts           = [];
-		$atts['title']  = ! empty( $item->attr_title ) ? $item->attr_title : '';
-		$atts['target'] = ! empty( $item->target ) ? $item->target : '';
-		if ( '_blank' === $item->target && empty( $item->xfn ) ) {
+		$atts['title']  = ! empty( $data_object->attr_title ) ? $menu_item->attr_title : '';
+		$atts['target'] = ! empty( $data_object->target ) ? $menu_item->target : '';
+		if ( '_blank' === $menu_item->target && empty( $menu_item->xfn ) ) {
 			$atts['rel'] = 'noopener';
 		} else {
-			$atts['rel'] = $item->xfn;
+			$atts['rel'] = $menu_item->xfn;
 		}
-		$atts['href']         = ! empty( $item->url ) ? $item->url : '';
-		$atts['aria-current'] = $item->current ? 'page' : '';
+		$atts['href']         = ! empty( $menu_item->url ) ? $menu_item->url : '';
+		$atts['aria-current'] = $menu_item->current ? 'page' : '';
 
 		/**
 		 * Filters the HTML attributes applied to a menu item's anchor element.
-		 *
-		 * @since 3.6.0
-		 * @since 4.1.0 The `$depth` parameter was added.
 		 *
 		 * @param array $atts {
 		 *     The HTML attributes applied to the menu item's `<a>` element, empty strings are ignored.
@@ -136,11 +135,11 @@ class PrimaryWalker extends \Walker_Nav_Menu {
 		 *     @type string $href         The href attribute.
 		 *     @type string $aria_current The aria-current attribute.
 		 * }
-		 * @param WP_Post  $item  The current menu item.
-		 * @param stdClass $args  An object of wp_nav_menu() arguments.
-		 * @param int      $depth Depth of menu item. Used for padding.
+		 * @param WP_Post  $menu_item The current menu item.
+		 * @param stdClass $args      An object of wp_nav_menu() arguments.
+		 * @param int      $depth     Depth of menu item. Used for padding.
 		 */
-		$atts = apply_filters( 'nav_menu_link_attributes', $atts, $item, $args, $depth );
+		$atts = apply_filters( 'nav_menu_link_attributes', $atts, $menu_item, $args, $depth );
 
 		$attributes = '';
 
@@ -160,28 +159,26 @@ class PrimaryWalker extends \Walker_Nav_Menu {
 		}
 
 		/** This filter is documented in wp-includes/post-template.php */
-		$title = apply_filters( 'the_title', $item->title, $item->ID );
+		$title = apply_filters( 'the_title', $menu_item->title, $menu_item->ID );
 
 		/**
 		 * Filters a menu item's title.
 		 *
-		 * @since 4.4.0
-		 *
-		 * @param string   $title The menu item's title.
-		 * @param WP_Post  $item  The current menu item.
-		 * @param stdClass $args  An object of wp_nav_menu() arguments.
-		 * @param int      $depth Depth of menu item. Used for padding.
+		 * @param string   $title     The menu item's title.
+		 * @param WP_Post  $menu_item The current menu item.
+		 * @param stdClass $args      An object of wp_nav_menu() arguments.
+		 * @param int      $depth     Depth of menu item. Used for padding.
 		 */
-		$title = apply_filters( 'nav_menu_item_title', $title, $item, $args, $depth );
+		$title = apply_filters( 'nav_menu_item_title', $title, $menu_item, $args, $depth );
 
 		$item_output  = $args->before;
 
 		if ( $args->walker && $args->walker->has_children ) {
 			$item_output .= '<button' . $attributes;
-			$item_output .= " @click.prevent='toggleSubMenu(" . $item->ID . ', ' . $item->menu_item_parent . ")'";
-			$item_output .= " @pointerenter.debounce.50ms='onPointerEnter(" . $item->ID . ', ' . $item->menu_item_parent . ")'";
+			$item_output .= " @click.prevent='toggleSubMenu(" . $menu_item->ID . ', ' . $menu_item->menu_item_parent . ")'";
+			$item_output .= " @pointerenter.debounce.50ms='onPointerEnter(" . $menu_item->ID . ', ' . $menu_item->menu_item_parent . ")'";
 			$item_output .= " @click.away='onClickAway'";
-			$item_output .= " :aria-expanded='(isSubMenuOpen(" . $item->ID . ")).toString()'";
+			$item_output .= " :aria-expanded='(isSubMenuOpen(" . $menu_item->ID . ")).toString()'";
 			$item_output .= " aria-haspopup='true'";
 			$item_output .= '>';
 			$item_output .= $args->link_before . $title . $args->link_after;
@@ -208,13 +205,11 @@ class PrimaryWalker extends \Walker_Nav_Menu {
 		 * the menu item's title, the closing `</a>`, and `$args->after`. Currently, there is
 		 * no filter for modifying the opening and closing `<li>` for a menu item.
 		 *
-		 * @since 3.0.0
-		 *
 		 * @param string   $item_output The menu item's starting HTML output.
-		 * @param WP_Post  $item        Menu item data object.
+		 * @param WP_Post  $menu_item   Menu item data object.
 		 * @param int      $depth       Depth of menu item. Used for padding.
 		 * @param stdClass $args        An object of wp_nav_menu() arguments.
 		 */
-		$output .= apply_filters( 'walker_nav_menu_start_el', $item_output, $item, $depth, $args );
+		$output .= apply_filters( 'walker_nav_menu_start_el', $item_output, $menu_item, $depth, $args );
 	}
 }
