@@ -19,7 +19,7 @@ class Setup implements Bootable {
 	 *
 	 * @return void
 	 */
-	public function boot() {
+	public function boot() : void {
 		add_action( 'after_setup_theme', [ $this, 'supports' ], 5 );
 		add_action( 'init', [ $this, 'menus' ] );
 		add_action( 'init', [ $this, 'image_sizes' ] );
@@ -32,7 +32,7 @@ class Setup implements Bootable {
 	 *
 	 * @return void
 	 */
-	public function supports() {
+	public function supports() : void {
 
 		// Theme translations.
 		load_theme_textdomain( 'pulsar', get_parent_theme_file_path( 'languages' ) );
@@ -80,7 +80,7 @@ class Setup implements Bootable {
 	 *
 	 * @return void
 	 */
-	public function menus() {
+	public function menus() : void {
 
 		register_nav_menus(
 			[
@@ -95,7 +95,7 @@ class Setup implements Bootable {
 	 *
 	 * @return void
 	 */
-	public function image_sizes() {
+	public function image_sizes() : void {
 		// add_image_size( '4x3', 640, 480, true );
 	}
 
@@ -106,7 +106,7 @@ class Setup implements Bootable {
 	 *
 	 * @return array
 	 */
-	public function image_size_names( $sizes ) {
+	public function image_size_names( array $sizes ) : array {
 		return array_merge(
 			$sizes,
 			[
@@ -122,8 +122,7 @@ class Setup implements Bootable {
 	 *
 	 * @return void
 	 */
-	public function javascript_detected() {
-
+	public function javascript_detected() : void {
 		echo "<script>(function(html){html.className = html.className.replace(/\bno-js\b/,'js')})(document.documentElement);</script>\n";
 	}
 }
